draft split Settings component
