+++ src/components/tabs/Settings/IntervalInput.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const IntervalInput = () => {
+  return <p>1. Link mobile phone to your account</p>;
+};
+
+export default IntervalInput;
